// src/api/generated/model/invitationType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

export type InvitationType = typeof InvitationType[keyof typeof InvitationType];


export const InvitationType = {
  REQUEST: 'REQUEST',
  INVITE: 'INVITE',
} as const;